DUPLICATE_CRAP_FROM_RESEARCH_WORKSPACE/Claude_SDK_Development/pyproject.toml
[tool.poetry]
name = "research-analytics-services"
version = "0.1.0"
description = "Research and Analytics Services with FRED data collection"
authors = ["HEAD OF DATA ENGINEERING"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.4"
pandas = "^2.3.0"
numpy = "^2.3.0"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
plotly = "^6.1.2"
azure-storage-blob = "^12.25.1"
azure-identity = "^1.23.0"
azure-keyvault-secrets = "^4.9.0"
python-dotenv = "^1.1.0"
tenacity = "^9.1.2"
redis = "^6.2.0"
fredapi = "^0.5.1"
ratelimit = "^2.2.1"
# Ultimate Azure Tooling
azure-mgmt-containerinstance = "^10.1.0"
azure-monitor-query = "^1.2.0"
azure-mgmt-redis = "^14.3.0"
azure-mgmt-storage = "^21.1.0"
azure-mgmt-monitor = "^6.0.0"
azure-mgmt-resource = "^23.0.0"
azure-mgmt-keyvault = "^10.3.0"
azure-eventhub = "^5.12.0"
azure-functions = "^1.20.0"
azure-cosmos = "^4.5.0"
azure-data-tables = "^12.5.0"
azure-servicebus = "^7.12.0"
azure-mgmt-compute = "^31.0.0"
azure-mgmt-network = "^26.0.0"
azure-applicationinsights = "^0.1.0"
azure-synapse-artifacts = "^0.20.0"
azure-synapse-spark = "^0.7.0"
azure-synapse-monitoring = "^0.2.0"
pyodbc = "^5.0.0"
pymssql = "^2.3.1"
anthropic = "^0.54.0"
gremlinpython = "^3.7.3"
scikit-learn = "^1.7.0"
networkx = "^3.5"
pyvis = "^0.3.2"
xgboost = "^3.0.2"
prophet = "^1.1.7"
statsmodels = "^0.14.4"
torch = "^2.7.1"
arch = "^7.2.0"
sentence-transformers = "^4.1.0"
pulp = "^3.2.1"
dask = "^2025.5.1"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
yfinance = "^0.2.63"
backtrader = "^1.9.78.123"
ta-lib = "^0.6.4"
mcp = "^1.10.1"
structlog = "^25.4.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
claude-code-sdk = "^0.0.14"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.16.0"
ipython = "^9.3.0"
jupyter = "^1.1.1"
nbconvert = "^7.16.6"

